requirements.txt
pandas
pyarrow
fastparquet
schedule
apache-airflow
streamlit